move validity checks from apply
